feat: consolidate skill render implementations
Remove duplicate renderSkill function from skill.ts and consolidate
into renderClaudeCodeSkill in skill-render.ts.

Changes:
- Move hash/drift functions to skill-render.ts and export them:
  computeContentHash, getRenderHashPath, readRenderHash, writeRenderHash,
  checkSkillDrift
- Add storeHash option to renderClaudeCodeSkill for drift detection
- Update CLI skill render command to use consolidated renderClaudeCodeSkill
- Remove ~200 lines of duplicate code from skill.ts

The CLI now passes storeHash: true to enable drift detection, while
setup.ts continues to use the default (no hash storage) for one-time
setup operations.

Task: @implement-consolidate-skill-render-implementations
Spec: @consolidate-skill-render
